+++ apple-codesign/src/macho.rs
@@ -64,7 +64,7 @@ pub enum CodeSigningSlot {
     ResourceDir,
     Application,
     Entitlements,
-    SecuritySettings,
+    EntitlementsDer,
     AlternateCodeDirectory0,
     AlternateCodeDirectory1,
     AlternateCodeDirectory2,
@@ -89,8 +89,8 @@ impl std::fmt::Debug for CodeSigningSlot {
             Self::ResourceDir => f.write_fmt(format_args!("Resources ({})", u32::from(*self))),
             Self::Application => f.write_fmt(format_args!("Application ({})", u32::from(*self))),
             Self::Entitlements => f.write_fmt(format_args!("Entitlements ({})", u32::from(*self))),
-            Self::SecuritySettings => {
-                f.write_fmt(format_args!("SecuritySettings ({})", u32::from(*self)))
+            Self::EntitlementsDer => {
+                f.write_fmt(format_args!("DER Entitlements ({})", u32::from(*self)))
             }
             Self::AlternateCodeDirectory0 => f.write_fmt(format_args!(
                 "CodeDirectory Alternate #0 ({})",
@@ -131,7 +131,7 @@ impl From<u32> for CodeSigningSlot {
             3 => Self::ResourceDir,
             4 => Self::Application,
             5 => Self::Entitlements,
-            7 => Self::SecuritySettings,
+            7 => Self::EntitlementsDer,
             0x1000 => Self::AlternateCodeDirectory0,
             0x1001 => Self::AlternateCodeDirectory1,
             0x1002 => Self::AlternateCodeDirectory2,
@@ -154,7 +154,7 @@ impl From<CodeSigningSlot> for u32 {
             CodeSigningSlot::ResourceDir => 3,
             CodeSigningSlot::Application => 4,
             CodeSigningSlot::Entitlements => 5,
-            CodeSigningSlot::SecuritySettings => 7,
+            CodeSigningSlot::EntitlementsDer => 7,
             CodeSigningSlot::AlternateCodeDirectory0 => 0x1000,
             CodeSigningSlot::AlternateCodeDirectory1 => 0x1001,
             CodeSigningSlot::AlternateCodeDirectory2 => 0x1002,